pyproject.toml
[project]
name = "pdf-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pdfminer-six>=20240706",
    "requests>=2.32.5",
]

[project.scripts]
pdf-tool = "pdf_tool.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
